curl: move benches to separate folder curl: add inlining annotations
