--- curl/src/bench.rs
@@ -1,129 +1,0 @@
-use test::Bencher;
-
-use collections::Vec;
-use trytes::*;
-use curl::*;
-
-#[bench]
-fn curl_pair(b: &mut Bencher) {
-    let trans: Trinary = "9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                999999999999999999999999999999T999999999999999999999999999999\
-                99999999999999999999999OLOB99999999999999999999999"
-        .chars()
-        .collect();
-
-    let trits : Vec<BCTrit> = trans.trits();
-    let mut curl = Curl::<BCTrit>::default();
-    curl.reset();
-    curl.absorb(&trits);
-    curl.squeeze(243);
-
-    b.iter(|| {
-        curl.reset();
-        curl.absorb(&trits);
-        curl.squeeze(243);
-    })
-}
-
-#[bench]
-fn curl_simple(b: &mut Bencher) {
-    let trans: Trinary = "9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                9999999999999999999999999999999999999999999999999999999999999\
-                999999999999999999999999999999T999999999999999999999999999999\
-                99999999999999999999999OLOB99999999999999999999999"
-        .chars()
-        .collect();
-
-    let trits : Vec<Trit> = trans.trits();
-    let mut curl = Curl::<Trit>::default();
-    curl.reset();
-    curl.absorb(&trits);
-    curl.squeeze(243);
-
-    b.iter(|| {
-        curl.reset();
-        curl.absorb(&trits);
-        curl.squeeze(243);
-    })
-}
