+++ curl/src/lib.rs
@@ -1,12 +1,7 @@
-#![allow(unused_features)]
-#![feature(test)]
 #![feature(alloc)]
 #![no_std]
 extern crate alloc as collections;
 
-
-#[cfg(test)]
-extern crate test;
 
 #[cfg(feature = "parallel")]
 extern crate rayon;
@@ -18,9 +13,6 @@ pub mod curl;
 pub mod simple;
 pub mod pair;
 mod tests;
-
-#[cfg(test)]
-mod bench;
 
 pub use tests::testsuite;
 pub use curl::*;
